Remove PVPREIFX restriction Merge tests and move source into appropriate directories.
